requirements.txt
pymongo==3.8.0
PyMySQL==0.7.11
Scrapy==1.7.3
scrapy-redis==0.6.8
fake-useragent==0.1.11